+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning].
     - `GET /echo` : returns the request data structure
     - `POST /upload` : multipart upload files
     - `GET /download`: download a file
+  - Command Line :
+    - `-n` or ``--no-browser` : when set, default browser is *not* opened on startup
 
 <!-- Links -->
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/